requirements.txt
aioboto3==13.1.1
aiobotocore==2.13.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioitertools==0.11.0
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
bcrypt==3.2.2
boltons==24.0.0
boto3==1.34.131
botocore==1.34.131
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
environs==11.0.0
fastapi==0.110.2
frozenlist==1.4.1
furl==2.1.3
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==2.10
iniconfig==2.0.0
iso8601==2.1.0
jmespath==1.0.1
libcst==1.4.0
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
MonkeyType==23.3.0
multidict==6.0.5
mypy-extensions==1.0.0
orderedmultidict==1.0.1
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
propcache==0.2.0
protobuf==4.25.5
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pypika-tortoise==0.2.2
pytest==8.3.2
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
ruff==0.6.4
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tortoise-orm==0.21.7
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
wrapt==1.16.0
yarl==1.17.2
